+++ src/App.jsx
@@ -310,6 +310,8 @@ class App extends AppBase {
               resumeAutoChoice={this.state.resumeAutoChoice}
               onResumeAutoChoiceToggle={this.handleResumeAutoChoiceToggle}
               onResumeAutoChoiceChange={this.handleResumeAutoChoiceChange}
+              themeColor={this.state.themeColor}
+              onThemeColorChange={this.handleThemeColorChange}
               proxyProfiles={this.state.proxyProfiles}
               activeProxyId={this.state.activeProxyId}
               defaultConfig={this.state.defaultConfig}